Initially split out the file processor into the processing of pupil /teacher etc., and a separate processor fro the leson model
