+++ timetable_solutions/domain/data_upload_processing/__init__.py
@@ -3,5 +3,6 @@ Convenience imports. Within the interfaces layer we simply import data_upload_pr
 """
 from .constants import FileStructure, UploadFileStructure
 from .file_upload_processor import FileUploadProcessor
+from .lesson_file_upload_processor import LessonFileUploadProcessor
 from .upload_status_tracking import UploadStatusTracker, UploadStatus
 from .reset_uploads import ResetUploads
